infra/pyproject.toml
[tool.poetry]
name = "infra"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.6"
"aws-cdk.core" = "^1.122.0"
"aws-cdk.aws-dynamodb" = "^1.122.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
